+++ releasenotes.md
@@ -1,5 +1,8 @@
 Whats New in v10.0
 --------------------------
+- Added inventory create/update info for compounds
+- Added JSON import/export for compounds
+- Allow a formula to be marked complete when contains skipped materials
 - Bottle edit page format update
 - Weight added for the bottles inventory
 - Added Inventory for finished Compounds